test(search): colocate UDF regression tests
